@@ -39,7 +39,7 @@ sealed class DisposingAggregateSink : ILogEventSink, IDisposable
             catch (Exception ex)
             {
                 SelfLog.WriteLine("Caught exception while emitting to sink {0}: {1}", sink, ex);
-                exceptions ??= new();
+                exceptions ??= [];
                 exceptions.Add(ex);
             }
         }
