@@ -8,7 +8,7 @@ public class MessageTemplateTests
     {
         // ReSharper disable UnusedMember.Local
         public string Back => "straight";
-        public int[] Legs => new[] { 1, 2, 3, 4 };
+        public int[] Legs => [1, 2, 3, 4];
         // ReSharper restore UnusedMember.Local
         public override string ToString() => "a chair";
     }
