@@ -9,7 +9,7 @@ public class DictionaryValueTests
             new KeyValuePair<ScalarValue, LogEventPropertyValue>(
                 new ScalarValue(1), new ScalarValue("hello")),
             new KeyValuePair<ScalarValue, LogEventPropertyValue>(
-                new ScalarValue("world"), new SequenceValue(new [] { new ScalarValue(1.2) }))
+                new ScalarValue("world"), new SequenceValue([new ScalarValue(1.2)]))
         });
 
         var sw = new StringWriter();
