@@ -2,7 +2,7 @@ namespace Serilog.Tests.Support;
 
 class CollectingSink : ILogEventSink
 {
-    public List<LogEvent> Events { get; } = new();
+    public List<LogEvent> Events { get; } = [];
 
     public LogEvent SingleEvent => Events.Single();
 
