@@ -5,7 +5,7 @@ public class DummyRollingFileAuditSink : ILogEventSink
     [ThreadStatic]
     static List<LogEvent>? _emitted;
 
-    public static List<LogEvent> Emitted => _emitted ??= new();
+    public static List<LogEvent> Emitted => _emitted ??= [];
 
     public void Emit(LogEvent logEvent)
     {
