@@ -5,7 +5,7 @@ public class DummyWrappingSink : ILogEventSink
     [ThreadStatic]
     static List<LogEvent>? _emitted;
 
-    public static List<LogEvent> Emitted => _emitted ??= new();
+    public static List<LogEvent> Emitted => _emitted ??= [];
 
     readonly ILogEventSink _sink;
 
